Add is_dependent to get_all_object_owners() results
In a subsequent commit we will be combining the
get_all_non_schema_objects_and_owners() function with
get_all_object_owners(). This commit is preparation for that by taking
the one component missing from get_all_object_owners() -- the
is_dependent field -- and adding it in.

This change means that get_all_object_owners() now has a dict at its
lowest level, i.e. for a given object (e.g. myschema.mytable) we now
have the object's owner as well as whether it is dependent. This opens
us up to expanding this function to holding other attributes of objects.
While it isn't clear yet whether there are other attributes we will want
to track, it's nice that this should be easy to accommodate if it proves
necessary.